+++ deepchem/models/__init__.py
@@ -12,13 +12,13 @@ from deepchem.models.multitask import SingletaskToMultitask
 from deepchem.models.callbacks import ValidationCallback
 
 from deepchem.models.tensorgraph.tensor_graph import TensorGraph
-from deepchem.models.tensorgraph.fcnet import MultitaskRegressor
-from deepchem.models.tensorgraph.fcnet import MultitaskClassifier
-from deepchem.models.tensorgraph.fcnet import MultitaskFitTransformRegressor
+from deepchem.models.fcnet import MultitaskRegressor
+from deepchem.models.fcnet import MultitaskClassifier
+from deepchem.models.fcnet import MultitaskFitTransformRegressor
 from deepchem.models.tensorgraph.IRV import TensorflowMultitaskIRVClassifier
-from deepchem.models.tensorgraph.robust_multitask import RobustMultitaskClassifier
-from deepchem.models.tensorgraph.robust_multitask import RobustMultitaskRegressor
-from deepchem.models.tensorgraph.progressive_multitask import ProgressiveMultitaskRegressor, ProgressiveMultitaskClassifier
+from deepchem.models.robust_multitask import RobustMultitaskClassifier
+from deepchem.models.robust_multitask import RobustMultitaskRegressor
+from deepchem.models.progressive_multitask import ProgressiveMultitaskRegressor, ProgressiveMultitaskClassifier
 from deepchem.models.tensorgraph.models.graph_models import WeaveModel, DTNNModel, DAGModel, GraphConvModel, MPNNModel
 from deepchem.models.tensorgraph.models.symmetry_function_regression import BPSymmetryFunctionRegression, ANIRegression
 from deepchem.models.tensorgraph.models.scscore import ScScoreModel
